+++ src/opts.h
@@ -15,6 +15,8 @@ extern "C" {
 //     unsigned int so_rcvtimeo;
 //     unsigned int tcp_user_timeout;
 //     unsigned int list_limit;
+//     unsigned int write_throttle_limit;
+//     unsigned int write_backlog_capacity;
 // };
 
 int rawstor_opts_initialize(const struct RawstorOpts* opts);
@@ -32,6 +34,10 @@ unsigned int rawstor_opts_so_rcvtimeo(void);
 unsigned int rawstor_opts_tcp_user_timeout(void);
 
 unsigned int rawstor_opts_list_limit(void);
+
+unsigned int rawstor_opts_write_throttle_limit(void);
+
+unsigned int rawstor_opts_write_backlog_capacity(void);
 
 #ifdef __cplusplus
 }